Added Post-Processing options
Added copy/move/symlink/hardlink options to the postprocessor. I’ve
done this because some torrent providers demand that you seed for X
days or until a certain ratio is hit. Now the processor can make a
hardlink/symlink so the file stays available in the torrent client as
well as the tvshows folder.  Default option is ‘Move’ as it was before.